Move Uuid => CoderSdk
