+++ Vpn.Proto/Vpn.Proto.csproj
@@ -21,4 +21,8 @@
         </PackageReference>
     </ItemGroup>
 
+    <ItemGroup>
+      <ProjectReference Include="..\CoderSdk\CoderSdk.csproj" />
+    </ItemGroup>
+
 </Project>
